// ui/src/app/core/api-client/model/file-entry.ts
/**
 * API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FileEntry { 
    name?: string;
    type?: FileEntry.TypeEnum;
    sizeBytes?: number;
    lastModified?: string;
}
export namespace FileEntry {
    export const TypeEnum = {
        File: 'FILE',
        Directory: 'DIRECTORY',
        Other: 'OTHER'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}